requirements.txt
anyio==3.7.0
asgiref==3.7.2
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.14
dnspython==2.3.0
fastapi==0.97.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.1
mongoengine==0.27.0
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-aws-lambda==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-tortoiseorm==0.39b0
opentelemetry-instrumentation-urllib==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
protobuf==4.23.3
pydantic==1.10.9
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
zipp==3.15.0
